Add date, hashState to notices
