+++ utils/hashState.ts
@@ -1,0 +1,34 @@
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
+
+export const useHashState = <S extends string | null>(
+  initialState: S | (() => S)
+): [S, Dispatch<SetStateAction<S>>] => {
+  const [state, setState] = useState<S>(initialState);
+
+  useEffect(() => {
+    if (location.hash) {
+      setState(location.hash.replace(/\#/, '') as S);
+    }
+
+    const hashChangeHandler = () =>
+      setState(
+        (location.hash === '' ? null : location.hash.replace(/\#/, '')) as S
+      );
+
+    window.addEventListener('hashchange', hashChangeHandler);
+
+    return () => {
+      window.removeEventListener('hashchange', hashChangeHandler);
+    };
+  }, []);
+
+  useEffect(() => {
+    if (location.hash === state || (state === null && location.hash === '')) {
+      return;
+    }
+
+    location.hash = state === null ? '' : `${state}`;
+  }, [state]);
+
+  return [state, setState];
+};
